Executive Summary
The most important assets of a modern society, besides our children, is our Land. It’s where we live work and play. It represents the freedom of ownership and is most often the most costly and valuable investment in our lifetime.

We often take this land ownership for granted but it has deep roots in American history and represents a major shift in the wealth of nations and individuals. It has allowed Americans to create an economy based on stable capital ownership thus creating enormous wealth. This solid foundation of land ownership offers unlimited potential for creation, land development and new construction, all resulting in a solid asset class.

As digital technology advances and it reaps benefits in our everyday lives these resulting efficiencies lower the costs of transactions and deliver a transparent alternative to existing data management systems. Adoption of these new blockchain technologies will increase the speed of transactions for real estate and offer a new paradigm in tracking and transacting a once labor intensive function. The ability to streamline these transaction services and automate laborious activities will result in a more efficient and cost effective method. Our Parallel Ledger system creates an engine for capital growth and manages the governmental costs associated with this complex but critical asset class.